Implemented a secondary constructor that allows an arbitrary shoe size to be set.
A dealing shoe or dealer's shoe is a gaming device, mainly used in casinos, to hold multiple decks of playing cards. The shoe allows for more games to be played by reducing the time between shuffles and less chance of dealer cheating. In some games, such as blackjack (where card counting is a possibility), using multiple decks of cards can increase the house edge.